removed db module; added those functions to command files
