+++ src/lib.rs
@@ -1,5 +1,4 @@
 #[macro_use]
 extern crate log;
 
-pub mod db;
 pub mod util;
